기존 cli로 graph stream방식 외에 paas용 파일 생성
