+++ requirements_paas.txt
@@ -1,0 +1,7 @@
+langchain-openai>=0.2.0
+langchain-community>=0.3.0
+langchain>=0.3.0
+langgraph>=0.2.26
+httpx>=0.24.0
+python-dotenv>=1.0.0
+adxp_sdk>=0.1.0
